backend/requirements.txt
fastapi
tortoise-orm
uvicorn[standard]
Jinja2
fastapi-mail

